Refactored the list and listitem schemas to control the order in which they're registered:
