+++ contexts/WorkspaceContext.tsx
@@ -12,9 +12,7 @@ export const WorkspaceProvider = (props) => {
 
 	const prepWorkspace = async (eventId: string) => {
 		//get the event and set it as the current event
-		const eventRes = await axios.get(`/api/events/${eventId}`, {
-			params: { id: eventId },
-		});
+		const eventRes = await axios.get(`/api/events/${eventId}`);
 		setCurrentEvent(eventRes.data.data);
 	};
